refactor(conversation): move the conversation agent into its own folder
