+++ libs/agents/application/use-cases/conversation-agent.use-case.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 
-import { ConversationAgentProvider } from '@libs/agents/infrastructure/services/agents/conversationAgent';
+import { ConversationAgentProvider } from '@libs/agents/infrastructure/services/agents/conversation/conversationAgent';
 import { IUseCase } from '@libs/core/domain/interfaces/use-case.interface';
 import { OrganizationAndTeamData } from '@libs/core/infrastructure/config/types/general/organizationAndTeamData';
 import { SandboxInstance } from '@libs/sandbox/domain/contracts/sandbox.provider';
