+++ libs/agents/infrastructure/services/agents/conversation/conversation-outcome.spec.ts
