+++ libs/agents/infrastructure/services/agents/conversation/conversation-tool-audit.spec.ts
