+++ libs/agents/infrastructure/services/agents/conversation/write-authorization.spec.ts
