+++ libs/agents/infrastructure/services/agents/conversation/write-truth.policy.ts
