+++ libs/agents/modules/agents.module.ts
@@ -15,7 +15,7 @@ import {
 import { BusinessRulesValidationAgentUseCase } from '../application/use-cases/business-rules-validation-agent.use-case';
 import { ConversationAgentUseCase } from '../application/use-cases/conversation-agent.use-case';
 import { BusinessRulesValidationAgentProvider } from '../infrastructure/services/agents/business-rules-validation/businessRulesValidationAgent';
-import { ConversationAgentProvider } from '../infrastructure/services/agents/conversationAgent';
+import { ConversationAgentProvider } from '../infrastructure/services/agents/conversation/conversationAgent';
 import { SkillLoaderService } from '../skills/skill-loader.service';
 import { GenericSkillRunnerService } from '../skills/generic-skill-runner.service';
 import { CapabilityStrategyService } from '../skills/runtime/capability-strategy.service';
